requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
antlr-python-runtime=4.7.2=py39h6e9494a_1003
asgiref=3.4.1=pypi_0
bcrypt=3.2.0=pypi_0
brotli=1.0.9=h0d85af4_6
brotli-bin=1.0.9=h0d85af4_6
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h0d85af4_4
c-ares=1.18.1=h0d85af4_0
ca-certificates=2021.10.8=h033912b_0
cartopy=0.20.1=py39hb41de3a_5
certifi=2021.10.8=py39h6e9494a_1
cf-units=3.0.1=py39hc89836e_2
cffi=1.14.6=py39h2125817_0
cftime=1.5.1.1=py39hc89836e_1
chardet=4.0.0=py39hecd8cb5_1003
cloudpickle=2.0.0=pyhd8ed1ab_0
conda=4.11.0=py39h6e9494a_0
conda-package-handling=1.7.3=py39h9ed2024_1
cryptography=3.4.7=py39h2fd3fbb_0
curl=7.80.0=hca72f7f_0
cycler=0.11.0=pyhd8ed1ab_0
dask-core=2021.12.0=pyhd8ed1ab_0
django=4.0=pypi_0
expat=2.4.2=he49afe7_0
fonttools=4.28.5=py39h89e85a6_0
freetype=2.10.4=h4cff582_1
fsspec=2021.11.1=pyhd8ed1ab_0
geos=3.10.1=he49afe7_1
h5py=3.6.0=pypi_0
hdf4=4.2.15=hefd3b78_3
hdf5=1.12.1=nompi_h2f0ef1a_100
idna=2.10=pyhd3eb1b0_0
iris=3.1.0=pyhd8ed1ab_3
jbig=2.1=h0d85af4_2003
jinja2=3.0.3=pyhd8ed1ab_0
jpeg=9d=hbcb3906_0
jsons=1.6.0=pypi_0
kiwisolver=1.3.2=py39hf018cea_1
krb5=1.19.2=hcfbf3a7_0
lcms2=2.12=h577c468_0
lerc=2.2.1=h046ec9c_0
libblas=3.9.0=12_osx64_openblas
libbrotlicommon=1.0.9=h0d85af4_6
libbrotlidec=1.0.9=h0d85af4_6
libbrotlienc=1.0.9=h0d85af4_6
libcblas=3.9.0=12_osx64_openblas
libcurl=7.80.0=h6dfd666_0
libcxx=12.0.1=habf9029_1
libdeflate=1.7=h35c211d_5
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
liblapack=3.9.0=12_osx64_openblas
libnetcdf=4.8.1=nompi_h6609ca0_101
libnghttp2=1.46.0=ha29bfda_0
libopenblas=0.3.18=openmp_h3351f45_0
libpng=1.6.37=h7cec526_2
libssh2=1.10.0=h52ee1ee_2
libtiff=4.3.0=h1167814_1
libwebp-base=1.2.1=h0d85af4_0
libzip=1.8.0=h8b0c345_1
llvm-openmp=12.0.1=hda6cdc1_1
locket=0.2.0=py_2
lz4-c=1.9.3=he49afe7_1
mako=1.1.6=pypi_0
markupsafe=2.0.1=py39h89e85a6_1
matplotlib-base=3.5.1=py39hb07454d_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.2=h0a44026_1
netcdf4=1.5.8=nompi_py39he7d1c46_101
networkx=2.6.3=pypi_0
numpy=1.21.5=pypi_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1l=h0d85af4_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
paramiko=2.9.1=pypi_0
partd=1.2.0=pyhd8ed1ab_0
pillow=8.4.0=pypi_0
pip=21.1.3=py39hecd8cb5_0
plotly=5.5.0=pypi_0
proj=8.2.0=h1512c50_0
pycosat=0.6.3=py39h9ed2024_0
pycparser=2.20=py_2
pynacl=1.4.0=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=3.0.6=pyhd8ed1ab_0
pyproj=3.3.0=py39hf9f0841_0
pyshp=2.1.3=pyh44b312d_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.5=h88f2d9e_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=2.0.2=py39h89e85a6_1
python.app=3=py39h9ed2024_0
python_abi=3.9=2_cp39
pytz=2021.3=pypi_0
pyyaml=6.0=py39h89e85a6_3
readline=8.1=h9ed2024_0
requests=2.25.1=pyhd3eb1b0_0
ruamel_yaml=0.15.100=py39h9ed2024_0
scipy=1.7.3=py39h056f1c0_0
seaborn=0.11.2=pypi_0
setuptools=52.0.0=py39hecd8cb5_0
shapely=1.8.0=py39h14969ae_4
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
sqlparse=0.4.2=pypi_0
tenacity=8.0.1=pypi_0
tk=8.6.10=hb0a8c7a_0
toolz=0.11.2=pyhd8ed1ab_0
tqdm=4.61.2=pyhd3eb1b0_1
typish=1.9.3=pypi_0
tzdata=2021a=h52ac0ba_0
udunits2=2.2.27.27=h06ef574_3
urllib3=1.26.6=pyhd3eb1b0_1
wheel=0.36.2=pyhd3eb1b0_0
xxhash=0.8.0=h35c211d_3
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zlib=1.2.11=h1de35cc_3
zstd=1.5.0=hcb37349_1
